feat(domain): firestore 스튜디오의 class class 구현
