+++ Domains/hooks/Firestore/index.ts
@@ -3,4 +3,5 @@ import Enrollment from "./Collections/Enrollment";
 import Student from "./Collections/Student";
 import Studio from "./Collections/Studio";
 
+// eslint-disable-next-line import/no-anonymous-default-export
 export default { Class, Enrollment, Student, Studio };
